adding alerts for deleting

+++ src/components/CartItem/cart-item.component.tsx
@@ -1,10 +1,10 @@
-import {
-  Button, CardActions, CardContent,
-} from '@mui/material';
+import { CardContent } from '@mui/material';
+import Swal from 'sweetalert2';
 import { ICartItemProps } from './cart-item.types';
 import {
   CardItem, Title, SmallText, MiddleText, BoxFlex,
 } from './cart-item.styles.js';
+import { SButton } from '../../pages/HomePage/HomePage.styles.js';
 
 export function CartItem({ item }: ICartItemProps) {
   const handleDelete = (id: string | number) => {
@@ -35,18 +35,43 @@ export function CartItem({ item }: ICartItemProps) {
             </SmallText>
           </div>
         </BoxFlex>
-      </CardContent>
-      <CardActions>
-        <Button
-          size="small"
+        <SButton
           onClick={() => {
-            console.log(item.id);
-            handleDelete(item.id);
+            Swal.fire({
+              title: 'Are you sure?',
+              text: "You won't be able to revert this!",
+              icon: 'warning',
+              showCancelButton: true,
+              confirmButtonColor: '#011638',
+              cancelButtonColor: '#011638',
+              confirmButtonText: 'Yes, delete it!',
+            }).then((result) => {
+              try {
+                console.log(item.id);
+                handleDelete(item.id);
+                if (result.isConfirmed) {
+                  Swal.fire({
+                    position: 'top-end',
+                    icon: 'success',
+                    title: 'Item has been deleted',
+                    showConfirmButton: false,
+                    timer: 1500,
+                  });
+                }
+              } catch (e) {
+                console.log(e);
+                Swal.fire({
+                  icon: 'error',
+                  title: 'Oops...',
+                  text: 'Something went wrong!',
+                });
+              }
+            });
           }}
         >
           Delete
-        </Button>
-      </CardActions>
+        </SButton>
+      </CardContent>
     </CardItem>
   );
 }
